api_ia/requirements-api_ia.txt
pytest==8.0.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
typing-inspect==0.9.0
fastapi==0.109.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
pyodbc==5.0.1
mlflow==2.12.1
mlflow-skinny==2.14.3
pydantic==2.6.1
pydantic_core==2.16.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
openai==1.23.2
python-dotenv==1.0.1
plotly==5.22.0
uvicorn==0.27.1
azureml-mlflow==1.56.0
azure-ai-ml==1.18.0